# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for Return Zero music app.",
        "title": "Return Zero API",
        "contact": {},
        "version": "1.0"
    },
    "host": "returnzero.ru",
    "basePath": "/",
    "paths": {
        "/albums": {
            "get": {
                "description": "Get a list of albums with optional pagination filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "Get albums",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Offset (default: 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit (default: 10, max: 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of albums",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/delivery.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/delivery.Album"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid filters",
                        "schema": {
                            "$ref": "#/definitions/delivery.APIBadRequestErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/delivery.APIInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/artists": {
            "get": {
                "description": "Get a list of artists with optional pagination filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "artists"
                ],
                "summary": "Get artists",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Offset (default: 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit (default: 10, max: 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of artists",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/delivery.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/delivery.Artist"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid filters",
                        "schema": {
                            "$ref": "#/definitions/delivery.APIBadRequestErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/delivery.APIInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/artists/{id}": {
            "get": {
                "description": "Get detailed information about a specific artist by their ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "artists"
                ],
                "summary": "Get artist by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Artist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artist details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/delivery.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/delivery.ArtistDetailed"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid ID",
                        "schema": {
                            "$ref": "#/definitions/delivery.APIBadRequestErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/delivery.APIInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/artists/{id}/albums": {
            "get": {
                "description": "Get a list of albums for a specific artist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "Get albums by artist ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Artist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of albums",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/delivery.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/delivery.Album"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid artist ID",
                        "schema": {
                            "$ref": "#/definitions/delivery.APIBadRequestErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/delivery.APIInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "delivery.APIBadRequestErrorResponse": {
            "description": "API bad request error response structure",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Something went wrong"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "delivery.APIInternalServerErrorResponse": {
            "description": "API internal server error response structure",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Something went wrong"
                },
                "status": {
                    "type": "integer",
                    "example": 500
                }
            }
        },
        "delivery.APIResponse": {
            "description": "API response wrapper",
            "type": "object",
            "properties": {
                "body": {},
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "delivery.Album": {
            "description": "A music album entity",
            "type": "object",
            "properties": {
                "artist": {
                    "type": "string"
                },
                "artist_id": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "release_date": {
                    "type": "string",
                    "example": "2021-01-01"
                },
                "thumbnail_url": {
                    "type": "string",
                    "example": "https://example.com/album.jpg"
                },
                "title": {
                    "type": "string",
                    "example": "Anticyclone"
                },
                "type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/delivery.AlbumType"
                        }
                    ],
                    "example": "album"
                }
            }
        },
        "delivery.AlbumType": {
            "type": "string",
            "enum": [
                "album",
                "ep",
                "compilation",
                "single"
            ],
            "x-enum-varnames": [
                "AlbumTypeAlbum",
                "AlbumTypeEP",
                "AlbumTypeCompilation",
                "AlbumTypeSingle"
            ]
        },
        "delivery.Artist": {
            "description": "A music artist entity",
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Inabakumori is a Japanese artist"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "thumbnail_url": {
                    "type": "string",
                    "example": "https://example.com/artist.jpg"
                },
                "title": {
                    "type": "string",
                    "example": "Inabakumori"
                }
            }
        },
        "delivery.ArtistDetailed": {
            "description": "A detailed music artist entity",
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Inabakumori is a Japanese artist"
                },
                "favorites_count": {
                    "type": "integer",
                    "example": 1000
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "listeners_count": {
                    "type": "integer",
                    "example": 1000
                },
                "thumbnail_url": {
                    "type": "string",
                    "example": "https://example.com/artist.jpg"
                },
                "title": {
                    "type": "string",
                    "example": "Inabakumori"
                }
            }
        },
...
    }
}