# internal/pkg/artist/delivery/http/handler.go
package artist

import (
	"net/http"
	"strconv"

	"github.com/go-park-mail-ru/2025_1_Return_Zero/config"
	"github.com/go-park-mail-ru/2025_1_Return_Zero/internal/middleware"
	"github.com/go-park-mail-ru/2025_1_Return_Zero/internal/pkg/artist"
	"github.com/go-park-mail-ru/2025_1_Return_Zero/internal/pkg/helpers"
	deliveryModel "github.com/go-park-mail-ru/2025_1_Return_Zero/internal/pkg/model/delivery"
	usecaseModel "github.com/go-park-mail-ru/2025_1_Return_Zero/internal/pkg/model/usecase"
	"github.com/gorilla/mux"
	"go.uber.org/zap"
)

type ArtistHandler struct {
	usecase artist.Usecase
	cfg     *config.Config
}

func NewArtistHandler(usecase artist.Usecase, cfg *config.Config) *ArtistHandler {
	return &ArtistHandler{usecase: usecase, cfg: cfg}
}

// GetAllArtists godoc
// @Summary Get artists
// @Description Get a list of artists with optional pagination filters
// @Tags artists
// @Accept json
// @Produce json
// @Param offset query integer false "Offset (default: 0)"
// @Param limit query integer false "Limit (default: 10, max: 100)"
// @Success 200 {object} delivery.APIResponse{body=[]delivery.Artist} "List of artists"
// @Failure 400 {object} delivery.APIBadRequestErrorResponse "Bad request - invalid filters"
// @Failure 500 {object} delivery.APIInternalServerErrorResponse "Internal server error"
// @Router /artists [get]
func (h *ArtistHandler) GetAllArtists(w http.ResponseWriter, r *http.Request) {
	logger := middleware.LoggerFromContext(r.Context())
	pagination, err := helpers.GetPagination(r, &h.cfg.Pagination)
	if err != nil {
		logger.Error("failed to get pagination", zap.Error(err))
		helpers.WriteErrorResponse(w, http.StatusBadRequest, err.Error(), nil)
		return
	}

	usecaseArtists, err := h.usecase.GetAllArtists(&usecaseModel.ArtistFilters{
		Pagination: &usecaseModel.Pagination{
			Offset: pagination.Offset,
			Limit:  pagination.Limit,
		},
	})

	if err != nil {
		logger.Error("failed to get artists", zap.Error(err))
		helpers.WriteErrorResponse(w, http.StatusInternalServerError, err.Error(), nil)
		return
	}

	artists := make([]*deliveryModel.Artist, 0, len(usecaseArtists))
	for _, usecaseArtist := range usecaseArtists {
		artists = append(artists, &deliveryModel.Artist{
			ID:          usecaseArtist.ID,
			Title:       usecaseArtist.Title,
			Thumbnail:   usecaseArtist.Thumbnail,
			Description: usecaseArtist.Description,
		})
	}
	helpers.WriteSuccessResponse(w, http.StatusOK, artists, nil)
}

// GetArtistByID godoc
// @Summary Get artist by ID
// @Description Get detailed information about a specific artist by their ID
// @Tags artists
// @Accept json
// @Produce json
// @Param id path integer true "Artist ID"
// @Success 200 {object} delivery.APIResponse{body=delivery.ArtistDetailed} "Artist details"
// @Failure 400 {object} delivery.APIBadRequestErrorResponse "Bad request - invalid ID"
// @Failure 500 {object} delivery.APIInternalServerErrorResponse "Internal server error"
// @Router /artists/{id} [get]
func (h *ArtistHandler) GetArtistByID(w http.ResponseWriter, r *http.Request) {
	logger := middleware.LoggerFromContext(r.Context())

	vars := mux.Vars(r)
	idStr := vars["id"]
	id, err := strconv.ParseInt(idStr, 10, 64)
	if err != nil {
		logger.Error("failed to parse artist ID", zap.Error(err))
		helpers.WriteErrorResponse(w, http.StatusBadRequest, err.Error(), nil)
		return
	}

	usecaseArtist, err := h.usecase.GetArtistByID(id)
	if err != nil {
		logger.Error("failed to get artist", zap.Error(err))
		helpers.WriteErrorResponse(w, http.StatusInternalServerError, err.Error(), nil)
		return
	}

	artistDetailed := &deliveryModel.ArtistDetailed{
		Artist: deliveryModel.Artist{
			ID:          usecaseArtist.ID,
			Title:       usecaseArtist.Title,
			Thumbnail:   usecaseArtist.Thumbnail,
			Description: usecaseArtist.Description,
		},
		Listeners: usecaseArtist.Listeners,
		Favorites: usecaseArtist.Favorites,
	}
	helpers.WriteSuccessResponse(w, http.StatusOK, artistDetailed, nil)
}